Updated for CanayMod 1.0-RC-1
Also removed Canary Classic files.
Old region files will still load ofc as that has nothign to do with mod implementation.

Fixed a couple of casts

+++ src/main/java/net/playblack/cuboids/datasource/CuboidSerializer.java
@@ -1,17 +1,18 @@
 package net.playblack.cuboids.datasource;
 
+import java.util.ArrayList;
+import java.util.HashMap;
+
 import net.playblack.cuboids.blocks.CBlock;
 import net.playblack.cuboids.blocks.CItem;
 import net.playblack.cuboids.blocks.ChestBlock;
 import net.playblack.cuboids.blocks.SignBlock;
 import net.playblack.cuboids.selections.CuboidSelection;
 import net.playblack.mcutils.Vector;
-import java.util.ArrayList;
-import java.util.HashMap;
 
 /**
  * This class takes care of serializing a CuboidSelection to text
- * 
+ *
  * @author chris
  */
 public abstract class CuboidSerializer {
@@ -28,7 +29,7 @@ public abstract class CuboidSerializer {
 
     /**
      * Load and serialize a cuboid selection into pieces of text
-     * 
+     *
      * @param cuboid
      */
     public CuboidSerializer(CuboidSelection cuboid) {
@@ -39,7 +40,7 @@ public abstract class CuboidSerializer {
 
     /**
      * Turn an ArrayList of BaseItems into a String for saving
-     * 
+     *
      * @param items
      * @return
      */
@@ -53,7 +54,7 @@ public abstract class CuboidSerializer {
 
     /**
      * Turn a Sign Text array into a string for saving
-     * 
+     *
      * @param text
      * @return
      */
@@ -72,7 +73,7 @@ public abstract class CuboidSerializer {
 
     /**
      * Serialize the cuboid into pieces of Strings.
-     * 
+     *
      * @param cuboid
      */
     private void serializeBlockList(CuboidSelection cuboid) {
@@ -84,13 +85,13 @@ public abstract class CuboidSerializer {
                     .toString());
 
             // Do chest block serializing
-            if ((Short) b.getType() == 54) {
+            if (b.getType() == 54) {
                 contents.put(baseData.size() - 1,
                         serializeItemList(((ChestBlock) b).getItemList()));
 
             }
             // Do SignBlock serializing
-            else if ((Short) b.getType() == 63) {
+            else if (b.getType() == 63) {
                 contents.put(baseData.size() - 1,
                         serializeSignText(((SignBlock) b).getSignTextArray()));
             }
